move tests into module directly
